+++ CHANGELOG.md
@@ -16,6 +16,11 @@
   This could be helpfull during development to pull all repos on instance.
 - Added command `odood pre-commit update` that could be used to update pre-commit dependencies.
   Just an alias for normal `pre-commit autoupdate` that is run inside correct repo dir and within correct venv.
+- Added option `--workers` to `deploy` command
+
+### Changed
+
+- Clean up pip cache after deploy, when docker image is built.
 
 ---